pyproject.toml
[tool.poetry]
name = "aiogram-template"
version = "0.1.0"
description = ""
authors = ["Vasiliynemix <vasyanemix@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.3.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
black = "^24.1.1"
loguru = "^0.7.2"
redis = "^5.0.1"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
